Finished the bulk of Billing and Service Delivery.

+++ facilities/src/main/java/edu/depaul/g6/facilities/domain/ServiceCategory.java
@@ -11,5 +11,5 @@ public class ServiceCategory {
     @Id
     private String id;
     private String category;
-    private double tariff;
+    private Integer tariff;
 }
